Initial commit
All that it does right now is display the 25mm_Cube.stl file in the viewer. You can move it around and select triangles. Then click to fit a plane to it. Next step is to get the plane where it actually needs to be on the cube. Right now it is offset. Then work on object tree left pane, and importing meshes.